07/11/2019

+++ couponSystemSpring/src/main/java/com/couponSystem/serviceImpl/CompanyServiceImpl.java
@@ -9,18 +9,21 @@ import org.springframework.stereotype.Service;
 
 import com.couponSystem.CouponClientFacade;
 import com.couponSystem.DateUtils;
+import com.couponSystem.exeptions.CouponDoesNotExistsException;
 import com.couponSystem.exeptions.CouponExistsException;
 import com.couponSystem.exeptions.CouponNotAvailableException;
 import com.couponSystem.javabeans.ClientType;
 import com.couponSystem.javabeans.Company;
 import com.couponSystem.javabeans.Coupon;
 import com.couponSystem.javabeans.CouponType;
+import com.couponSystem.javabeans.Customer;
 import com.couponSystem.javabeans.Income;
 import com.couponSystem.javabeans.IncomeType;
 import com.couponSystem.repository.CompanyRepository;
 import com.couponSystem.repository.CouponRepository;
-import com.couponSystem.repository.IncomeRepository;
+import com.couponSystem.repository.CustomerRepository;
 import com.couponSystem.service.CompanyService;
+import com.couponSystem.service.IncomeService;
 
 @Service
 @Repository
@@ -32,10 +35,13 @@ public class CompanyServiceImpl implements CompanyService, CouponClientFacade {
 	@Autowired
 	private CouponRepository couponRepository;
 	@Autowired
-	private IncomeRepository incomeRepository;
+	private CustomerRepository customerRepository;
+	@Autowired
+	private IncomeService incomeService;
 
 	private Company company;
 
+	@Override
 	public Company getCompany() {
 		return this.company;
 	}
@@ -49,27 +55,20 @@ public class CompanyServiceImpl implements CompanyService, CouponClientFacade {
 	}
 
 	@Override
-	public void createCoupon(Coupon coupon) throws Exception {
+	public synchronized void createCoupon(Coupon coupon) throws Exception {
 
-		if (CouponAlreadyExists(coupon.getTitle()) != true) {
+		if (!CouponAlreadyExists(coupon.getTitle())) {
 			Company company = this.companyRepository.findById(this.company.getId());
-			System.out.println("CompanyServiceImpl.0000000000000000000000");
 			this.couponRepository.save(coupon);
-
 			company.getCoupons().add(coupon);
-			System.out.println(coupon.toString());
-			System.out.println("CompanyServiceImpl.1111111111111111111");
-
 			this.companyRepository.save(company);
-			System.out.println("CompanyServiceImpl.2222222222222");
-
 			Income income = new Income();
 			income.setAmount(100.0);
 			income.setClientId(company.getId());
 			income.setDescription(IncomeType.COMPANY_NEW_COUPON);
 			income.setDate(DateUtils.GetCurrentDate());
 			income.setName("Company " + company.getCompName());
-			this.incomeRepository.save(income);
+			this.incomeService.createIncome(income);
 		} else {
 			throw new CouponExistsException(
 					" coupon with name " + coupon.getTitle() + " already exist, please tryanother name");
@@ -79,31 +78,56 @@ public class CompanyServiceImpl implements CompanyService, CouponClientFacade {
 	@Override
 	public void removeCoupon(long id) throws Exception {
 		List<Coupon> coupons = getCompCoupons();
-
+		System.out.println("in removeCoupon metod ");
 		Coupon coupon = this.couponRepository.findById(id);
+		if (coupon == null) {
+			throw new CouponDoesNotExistsException(" coupon does not exist, please try another Coupon");
+		}
 		coupons.remove(coupon);
 		this.company.setCoupons(coupons);
-
+		List<Customer> customers = this.customerRepository.findAll();
+		for (Customer c : customers) {
+			System.out.println("12345678910");
+			if (c.getCoupons().contains(coupon)) {
+				List<Coupon> custCoupons = c.getCoupons();
+				custCoupons.remove(coupon);
+				c.setCoupons(custCoupons);
+			}
+		}
 		this.couponRepository.delete(coupon);
 
 	}
 
 	@Override
-	public void updateCoupon(Coupon coupon) throws Exception {
+	public synchronized void updateCoupon(Coupon coupon) throws Exception {
+		if (coupon == null) {
+			throw new CouponDoesNotExistsException(" coupon does not exist, please try another Coupon");
+		}
 		this.couponRepository.save(coupon);
+		System.out.println("111111111111111");
 		Income income = new Income();
 		income.setAmount(10.0);
 		income.setClientId(this.company.getId());
 		income.setDescription(IncomeType.COMPANY_UPDATE_COUPON);
 		income.setDate(DateUtils.GetCurrentDate());
 		income.setName("Company " + this.company.getCompName());
-		this.incomeRepository.save(income);
+		this.incomeService.createIncome(income);
 	}
 
 	@Override
 	public Coupon getCoupon(long id) throws Exception {
-		Coupon coupon = this.couponRepository.findOne(id);
-		return coupon;
+
+		Company company = this.companyRepository.getOne(this.company.getId());
+		if (company.getCoupons().contains(this.couponRepository.findById(id))) {
+			Coupon coupon = this.couponRepository.findById(id);
+			if (coupon == null) {
+				throw new CouponDoesNotExistsException(" coupon does not exist, please try another Coupon");
+			}
+			return coupon;
+		} else {
+			throw new Exception("This company doesn't have such coupon");
+		}
+		// TODO: handle exception
 	}
 
 	@Override
@@ -124,15 +148,37 @@ public class CompanyServiceImpl implements CompanyService, CouponClientFacade {
 	@Override
 	public List<Coupon> getCouponsByType(CouponType couponType) throws Exception {
 		List<Coupon> compCoupons = getCompCoupons();
-		List<Coupon> coupons = new ArrayList<>();
+		List<Coupon> coupons = null;
 		System.out.println(compCoupons.toString());
 		for (Coupon c : compCoupons) {
 			System.out.println("12345678910");
 			if (c.getCouponType() == couponType) {
-
+				coupons = new ArrayList<>();
 				coupons.add(c);
 				System.out.println(c.toString());
 			}
+			if (coupons != null) {
+				throw new CouponNotAvailableException("This company doesn't have coupons of this type");
+			}
+		}
+		return coupons;
+	}
+
+	@Override
+	public List<Coupon> getCouponsByPrice(double price) throws Exception {
+		List<Coupon> compCoupons = getCompCoupons();
+		System.out.println(compCoupons.toString());
+		List<Coupon> coupons = null;
+		for (Coupon c : compCoupons) {
+			System.out.println("12345678910");
+			if (c.getPrice() <= price) {
+				coupons = new ArrayList<>();
+				coupons.add(c);
+				System.out.println(c.toString());
+			}
+		}
+		if (coupons == null) {
+			throw new CouponNotAvailableException("This company doesn't have coupons of this price");
 		}
 		return coupons;
 	}
@@ -144,7 +190,6 @@ public class CompanyServiceImpl implements CompanyService, CouponClientFacade {
 		return false;
 	}
 
-	@Override
 	public boolean loginCheck(String compName, String password) throws Exception {
 		boolean loginStatus = false;
 		try {
@@ -167,19 +212,4 @@ public class CompanyServiceImpl implements CompanyService, CouponClientFacade {
 		return this.companyRepository.findByCompNameAndPassword(compName, password);
 	}
 
-	@Override
-	public List<Coupon> getCouponsByPrice(double price) throws Exception {
-		List<Coupon> compCoupons = getCompCoupons();
-		List<Coupon> coupons = new ArrayList<>();
-		System.out.println(compCoupons.toString());
-		for (Coupon c : compCoupons) {
-			System.out.println("12345678910");
-			if (c.getPrice() >= price) {
-
-				coupons.add(c);
-				System.out.println(c.toString());
-			}
-		}
-		return coupons;
-	}
 }
